+++ src/Service/BedrockRuntime/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## NOT RELEASED
 
+### Added
+
+- AWS api-change: Adding new BDD representation of endpoint ruleset
+- AWS api-change: Supporting Request Metadata for Invoke Model and Invoke Model with Response Stream
+
 ## 1.2.0
 
 ### Added